joint speed control

+++ tools/java/crcl4java/crcl4java-motoman/src/main/java/com/github/wshackle/crcl4java/motoman/motctrl/MP_SPEED.java
@@ -27,6 +27,6 @@ package com.github.wshackle.crcl4java.motoman.motctrl;
  * @author Will Shackleford {@literal <william.shackleford@nist.gov>}
  */
 public class MP_SPEED {
-    public long vj,v,vr;
+    public int vj,v,vr;
     
 }
